+++ tests/acceptance/fixtures/test_bashunit_when_setup_before_script_with_intermediate_failing_command.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+
+function set_up_before_script() {
+  false
+  true
+}
+
+function test_dummy() {
+  assert_same "foo" "foo"
+}