+++ tests/acceptance/fixtures/test_bashunit_when_teardown_after_script_with_intermediate_failing_command.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+
+function test_dummy() {
+  assert_same "foo" "foo"
+}
+
+function tear_down_after_script() {
+  false
+  true
+}